Some refactor and skeleton to save a new lot and its initial transaction Issue: IBP-3174 Reviewer: none
