+++ src/main/java/org/generationcp/middleware/service/api/inventory/LotService.java
@@ -1,5 +1,6 @@
 package org.generationcp.middleware.service.api.inventory;
 
+import org.generationcp.middleware.domain.inventory_new.ExtendedLotDto;
 import org.generationcp.middleware.domain.inventory_new.LotDto;
 import org.generationcp.middleware.domain.inventory_new.LotsSearchDto;
 import org.springframework.data.domain.Pageable;
@@ -8,8 +9,10 @@ import java.util.List;
 
 public interface LotService {
 
-	List<LotDto> searchLots(LotsSearchDto lotsSearchDto, Pageable pageable);
+	List<ExtendedLotDto> searchLots(LotsSearchDto lotsSearchDto, Pageable pageable);
 
 	long countSearchLots(LotsSearchDto lotsSearchDto);
+
+	Integer saveLot(LotDto lotDto);
 
 }
